Make direct children of loading-box always stack on top of each other
Practically it means that we should only put "component+its indicator" directly inside loading-box in order to ensure that their sizes and positioning always match (incl on dynamic resize while request is ongoing).
This logic  with some tweaks could be done by adding "loading-box" to the table directly (instead of adding a wrapper div) but there might be some styling conflicts due to "htmx-request" class being added some other places in the table.
Div wrapper serves both as a failsafe and reduces the complexity of style definitions in loading-box.css